feat: main페이지 스크롤 애니메이션 로직 분리
