+++ src/main.css
@@ -7,6 +7,7 @@
 
   /* color */
   --color-bg-black: #09090b;
+  --color-bg-black-100: #0e0f14;
   --color-bg-darkgray: #141620;
 
   --color-point-darkpurple: #404060;
@@ -64,6 +65,5 @@
 @layer base {
   html {
     font-family: var(--font-paperlogy);
-    scroll-behavior: smooth;
   }
 }
