+++ src/pages/ProjectPage.tsx
@@ -1,7 +1,7 @@
 const ProjectPage = () => {
   return (
-    <section className='h-screen'>
-      <h1>ProjectPage</h1>
+    <section className='bg-bg-black-100 h-screen'>
+      <h1 className='text-text-white'>ProjectPage</h1>
     </section>
   )
 }
